+++ autofixture/generators.py
@@ -1,10 +1,17 @@
 # -*- coding: utf-8 -*-
-from decimal import Decimal
 import datetime
+import os
 import random
 import re
 import string
-import os
+import sys
+from decimal import Decimal
+
+
+if sys.version_info[0] < 3:
+    str_ = unicode
+else:
+    str_ = str
 
 
 # backporting os.path.relpath, only availabe in python >= 2.6
@@ -82,8 +89,8 @@ class NoneGenerator(Generator):
 
 
 class StringGenerator(Generator):
-    coerce_type = unicode
-    singleline_chars = string.letters + u' '
+    coerce_type = str_
+    singleline_chars = string.ascii_letters + u' '
     multiline_chars = singleline_chars + u'\n'
 
     def __init__(self, chars=None, multiline=False, min_length=1, max_length=1000, *args, **kwargs):
@@ -103,7 +110,7 @@ class StringGenerator(Generator):
     def generate(self):
         length = random.randint(self.min_length, self.max_length)
         value = u''
-        for x in xrange(length):
+        for x in range(length):
             value += random.choice(self.chars)
         return value
 
@@ -116,7 +123,7 @@ class SlugGenerator(StringGenerator):
 
 
 class LoremGenerator(Generator):
-    coerce_type = unicode
+    coerce_type = str_
     common = True
     count = 3
     method = 'b'
@@ -138,14 +145,14 @@ class LoremGenerator(Generator):
             lorem = words(self.count, common=self.common)
         elif self.method == 's':
             lorem = u' '.join([sentence()
-                for i in xrange(self.count)])
+                for i in range(self.count)])
         else:
             paras = paragraphs(self.count, common=self.common)
             if self.method == 'p':
                 paras = ['<p>%s</p>' % p for p in paras]
             lorem = u'\n\n'.join(paras)
         if self.max_length:
-            length = random.randint(self.max_length / 10, self.max_length)
+            length = random.randint(round(self.max_length / 10), self.max_length)
             lorem = lorem[:max(1, length)]
         return lorem.strip()
 
@@ -237,14 +244,16 @@ class NullBooleanGenerator(BooleanGenerator):
 
 
 class DateTimeGenerator(Generator):
-    min_date = datetime.datetime.now() - datetime.timedelta(365 * 5)
-    max_date = datetime.datetime.now() + datetime.timedelta(365 * 1)
-
     def __init__(self, min_date=None, max_date=None, *args, **kwargs):
+        from django.utils import timezone
         if min_date is not None:
             self.min_date = min_date
+        else:
+            self.min_date = timezone.now() - datetime.timedelta(365 * 5)
         if max_date is not None:
             self.max_date = max_date
+        else:
+            self.max_date = timezone.now() + datetime.timedelta(365 * 1)
         assert self.min_date < self.max_date
         super(DateTimeGenerator, self).__init__(*args, **kwargs)
 
@@ -295,29 +304,83 @@ class DecimalGenerator(Generator):
         return value
 
 
+class FirstNameGenerator(Generator):
+    """ Generates a first name, either male or female """
+
+    male = [
+        'Abraham', 'Adam', 'Anthony', 'Brian', 'Bill', 'Ben', 'Calvin',
+        'David', 'Daniel', 'George', 'Henry', 'Isaac', 'Ian', 'Jonathan',
+        'Jeremy', 'Jacob', 'John', 'Jerry', 'Joseph', 'James', 'Larry',
+        'Michael', 'Mark', 'Paul', 'Peter', 'Phillip', 'Stephen', 'Tony',
+        'Titus', 'Trevor', 'Timothy', 'Victor', 'Vincent', 'Winston', 'Walt']
+    female = [
+        'Abbie', 'Anna', 'Alice', 'Beth', 'Carrie', 'Christina', 'Danielle',
+        'Emma', 'Emily', 'Esther', 'Felicia', 'Grace', 'Gloria', 'Helen',
+        'Irene', 'Joanne', 'Joyce', 'Jessica', 'Kathy', 'Katie', 'Kelly',
+        'Linda', 'Lydia', 'Mandy', 'Mary', 'Olivia', 'Priscilla',
+        'Rebecca', 'Rachel', 'Susan', 'Sarah', 'Stacey', 'Vivian']
+
+    def __init__(self, gender=None):
+        self.gender = gender
+        self.all = self.male + self.female
+
+    def generate(self):
+        if self.gender == 'm':
+            return random.choice(self.male)
+        elif self.gender == 'f':
+            return random.choice(self.female)
+        else:
+            return random.choice(self.all)
+
+
+class LastNameGenerator(Generator):
+    """ Generates a last name """
+    
+    surname = [
+        'Smith', 'Walker', 'Conroy', 'Stevens', 'Jones', 'Armstrong',
+        'Johnson', 'White', 'Stone', 'Strong', 'Olson', 'Lee', 'Forrest',
+        'Baker', 'Portman', 'Davis', 'Clark', 'Brown', 'Roberts', 'Ellis',
+        'Jackson', 'Marshall', 'Wang', 'Chen', 'Chou', 'Tang', 'Huang', 'Liu',
+        'Shih', 'Su', 'Song', 'Yang', 'Chan', 'Tsai', 'Wong', 'Hsu', 'Cheng',
+        'Chang', 'Wu', 'Lin', 'Yu', 'Yao', 'Kang', 'Park', 'Kim', 'Choi',
+        'Ahn', 'Mujuni']
+
+    def generate(self):
+        return random.choice(self.surname)
+
+
 class EmailGenerator(StringGenerator):
     chars = string.ascii_lowercase
 
-    def __init__(self, chars=None, max_length=30, tlds=None, *args, **kwargs):
+    def __init__(self, chars=None, max_length=30, tlds=None, static_domain=None, *args, **kwargs):
         assert max_length >= 6
         if chars is not None:
             self.chars = chars
         self.tlds = tlds
+        self.static_domain = static_domain
         super(EmailGenerator, self).__init__(self.chars, max_length=max_length, *args, **kwargs)
 
     def generate(self):
         maxl = self.max_length - 2
-        if self.tlds:
-            tld = random.choice(self.tlds)
-        elif maxl > 4:
-            tld = StringGenerator(self.chars, min_length=3, max_length=3).generate()
-        maxl -= len(tld)
-        assert maxl >= 2
+
+        if self.static_domain is None:
+            if self.tlds:
+                tld = random.choice(self.tlds)
+            elif maxl > 4:
+                tld = StringGenerator(self.chars, min_length=3, max_length=3).generate()
+            maxl -= len(tld)
+            assert maxl >= 2
+        else:
+            maxl -= len(self.static_domain)
 
         name = StringGenerator(self.chars, min_length=1, max_length=maxl-1).generate()
         maxl -= len(name)
-        domain = StringGenerator(self.chars, min_length=1, max_length=maxl).generate()
-        return '%s@%s.%s' % (name, domain, tld)
+
+        if self.static_domain is None:
+            domain = StringGenerator(self.chars, min_length=1, max_length=maxl).generate()
+            return '%s@%s.%s' % (name, domain, tld)
+        else:
+            return '%s@%s' % (name, self.static_domain)
 
 
 class URLGenerator(StringGenerator):
@@ -355,10 +418,10 @@ class URLGenerator(StringGenerator):
 
 
 class IPAddressGenerator(Generator):
-    coerce_type = unicode
+    coerce_type = str_
 
     def generate(self):
-        return '.'.join([unicode(part) for part in [
+        return '.'.join([str_(part) for part in [
             IntegerGenerator(min_value=1, max_value=254).generate(),
             IntegerGenerator(min_value=0, max_value=254).generate(),
             IntegerGenerator(min_value=0, max_value=254).generate(),
@@ -367,7 +430,7 @@ class IPAddressGenerator(Generator):
 
 
 class TimeGenerator(Generator):
-    coerce_type = unicode
+    coerce_type = str_
 
     def generate(self):
         return u'%02d:%02d:%02d' % (
@@ -378,7 +441,7 @@ class TimeGenerator(Generator):
 
 
 class FilePathGenerator(Generator):
-    coerce_type = unicode
+    coerce_type = str_
 
     def __init__(self, path, match=None, recursive=False, max_length=None, *args, **kwargs):
         self.path = path
@@ -462,7 +525,7 @@ class MultipleInstanceGenerator(InstanceGenerator):
 
     def generate(self):
         instances = []
-        for i in xrange(random.randint(self.min_count, self.max_count)):
+        for i in range(random.randint(self.min_count, self.max_count)):
             instances.append(
                 super(MultipleInstanceGenerator, self).generate())
         return instances
@@ -496,3 +559,26 @@ class InstanceSelector(Generator):
             min_count = self.min_count or 0
             count = random.randint(min_count, self.max_count)
             return self.queryset.order_by('?')[:count]
+
+class WeightedGenerator(Generator):
+    """ 
+    Takes a list of generator objects and integer weights, of the following form:
+    [(generator, weight), (generator, weight),...]
+    and returns a value from a generator chosen randomly by weight.
+    """
+
+    def __init__(self, choices):
+        self.choices = choices
+
+    def weighted_choice(self, choices):
+        total = sum(w for c, w in choices)
+        r = random.uniform(0, total)
+        upto = 0
+        for c, w in choices:
+          if upto + w > r:
+             return c
+          upto += w
+
+    def generate(self):
+        return self.weighted_choice(self.choices).generate()
+