+++ autofixture/management/commands/loadtestdata.py
@@ -25,6 +25,7 @@ information::
 
     django-admin.py help loadtestdata
 '''
+from django.utils.encoding import smart_text
 import autofixture
 from django.db import models
 from django.db.transaction import commit_on_success
@@ -82,41 +83,41 @@ class Command(BaseCommand):
     )
 
     def format_output(self, obj):
-        output = unicode(obj)
+        output = smart_text(obj)
         if len(output) > 50:
-            output = u'%s ...' % output[:50]
+            output = u'{0} ...'.format(output[:50])
         return output
 
     def print_instance(self, sender, model, instance, **kwargs):
         if self.verbosity < 1:
             return
-        print '%s(pk=%s): %s' % (
-            '%s.%s' % (
+        print('{0}(pk={1}): {2}'.format(
+            '.'.join((
                 model._meta.app_label,
-                model._meta.object_name),
-            unicode(instance.pk),
+                model._meta.object_name)),
+            smart_text(instance.pk),
             self.format_output(instance),
-        )
+        ))
         if self.verbosity < 2:
             return
         for field in instance._meta.fields:
             if isinstance(field, models.ForeignKey):
                 obj = getattr(instance, field.name)
                 if isinstance(obj, models.Model):
-                    print '|   %s (pk=%s): %s' % (
+                    print('|   {0} (pk={1}): {2}'.format(
                         field.name,
                         obj.pk,
-                        self.format_output(obj))
+                        self.format_output(obj)))
         for field in instance._meta.many_to_many:
             qs = getattr(instance, field.name).all()
             if qs.count():
-                print '|   %s (count=%d):' % (
+                print('|   {0} (count={1}):'.format(
                     field.name,
-                    qs.count())
+                    qs.count()))
                 for obj in qs:
-                    print '|   |   (pk=%s): %s' % (
+                    print('|   |   (pk={0}): {1}'.format(
                         obj.pk,
-                        self.format_output(obj))
+                        self.format_output(obj)))
 
     @commit_on_success
     def handle(self, *attrs, **options):
@@ -144,7 +145,7 @@ class Command(BaseCommand):
                             key, minval, maxval = field.split(':')
                             follow_m2m[key] = int(minval), int(maxval)
                 except ValueError:
-                    error_option = '--follow-m2m=%s' % options['follow_m2m']
+                    error_option = '--follow-m2m={0}'.format(options['follow_m2m'])
         else:
             follow_m2m = False
         #
@@ -166,12 +167,12 @@ class Command(BaseCommand):
                         key, minval, maxval = field.split(':')
                         generate_m2m[key] = int(minval), int(maxval)
             except ValueError:
-                error_option = '--generate-m2m=%s' % options['generate_m2m']
+                error_option = '--generate-m2m={0}'.format(options['generate_m2m'])
 
         if error_option:
             raise CommandError(
-                u'Invalid option %s\n'
-                u'Expected: %s=field:min:max,field2:min:max... (min and max must be numbers)' % (
+                u'Invalid option {0}\n'
+                u'Expected: {1}=field:min:max,field2:min:max... (min and max must be numbers)'.format(
                     error_option,
                     error_option.split('=', 1)[0]))
 
@@ -191,16 +192,13 @@ class Command(BaseCommand):
                 count = int(count)
             except ValueError:
                 raise CommandError(
-                    u'Invalid argument: %s\n'
+                    u'Invalid argument: {0}\n'
                     u'Expected: app_label.ModelName:count '
-                    u'(count must be a number)' % (
-                        attr,
-                    )
-                )
+                    u'(count must be a number)'.format(attr))
             model = get_model(app_label, model_label)
             if not model:
                 raise CommandError(
-                    u'Unknown model: %s.%s' % (app_label, model_label))
+                    u'Unknown model: {0}.{1}'.format(app_label, model_label))
             models.append((model, count))
 
         signals.instance_created.connect(