+++ autofixture_tests/tests/__init__.py
@@ -1,0 +1,5 @@
+from .autodiscover import *
+from .base import *
+from .generator import *
+from .user_fixture import *
+from .values import *