+++ autofixture_tests/tests/base.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import sys
 import autofixture
 from decimal import Decimal
 from datetime import date, datetime
@@ -6,11 +7,19 @@ from django.test import TestCase
 from autofixture import generators
 from autofixture.base import AutoFixture, CreateInstanceError,  Link
 from autofixture.values import Values
-from autofixture_tests.autofixture_test.models import y2k
-from autofixture_tests.autofixture_test.models import (
+from ..models import y2k
+from ..models import (
     SimpleModel, OtherSimpleModel, DeepLinkModel1, DeepLinkModel2,
     NullableFKModel, BasicModel, UniqueTestModel, UniqueTogetherTestModel,
-    RelatedModel, O2OModel, M2MModel, ThroughModel, M2MModelThrough) 
+    RelatedModel, O2OModel, InheritModel, InheritUniqueTogetherModel,
+    M2MModel, ThroughModel, M2MModelThrough, SelfReferencingModel,
+    SelfReferencingModelNoNull, GFKModel, GRModel)
+
+
+if sys.version_info[0] < 3:
+    str_ = unicode
+else:
+    str_ = str
 
 
 class SimpleAutoFixture(AutoFixture):
@@ -50,12 +59,12 @@ class TestBasicModel(TestCase):
             overwrite_defaults=False)
         for obj in filler.create(100):
             self.assertTrue(len(obj.chars) > 0)
-            self.assertEqual(type(obj.chars), unicode)
+            self.assertEqual(type(obj.chars), str_)
             self.assertTrue(len(obj.shortchars) <= 2)
-            self.assertEqual(type(obj.shortchars), unicode)
-            self.assertTrue(type(obj.blankchars), unicode)
-            self.assertEqualOr(type(obj.nullchars), unicode, None)
-            self.assertEqual(type(obj.slugfield), unicode)
+            self.assertEqual(type(obj.shortchars), str_)
+            self.assertTrue(type(obj.blankchars), str_)
+            self.assertEqualOr(type(obj.nullchars), str_, None)
+            self.assertEqual(type(obj.slugfield), str_)
             self.assertEqual(type(obj.defaultint), int)
             self.assertEqual(obj.defaultint, 1)
             self.assertEqual(type(obj.intfield), int)
@@ -251,6 +260,45 @@ class TestRelations(TestCase):
             all_m2m.update(obj.m2m.all())
         self.assertEqual(SimpleModel.objects.count(), len(all_m2m))
 
+    def test_generate_fk_to_self(self):
+        ''' When a model with a reference to itself is encountered, If NULL is allowed
+            don't generate a new instance of itself as a foreign key, so as not to reach
+            pythons recursion limit
+        '''
+        filler = AutoFixture(SelfReferencingModel, generate_fk=True)
+        model = filler.create_one()
+        self.assertEqual(model.parent_self, None)
+        self.assertEqual(SelfReferencingModel.objects.count(), 1)
+
+    def test_generate_fk_to_self_no_null(self):
+        ''' Throw an exception when a model is encountered which references itself but
+            does not allow NULL values to be set.
+        '''
+        filler = AutoFixture(SelfReferencingModelNoNull, generate_fk=True)
+        self.assertRaises(CreateInstanceError, filler.create_one)
+
+    def test_generate_fk_to_self_follow(self):
+        filler = AutoFixture(SelfReferencingModel, follow_fk=True)
+        first = filler.create_one()
+        self.assertEqual(SelfReferencingModel.objects.count(), 1)
+
+        filler = AutoFixture(SelfReferencingModel, follow_fk=True)
+        second = filler.create_one()
+        self.assertEqual(SelfReferencingModel.objects.count(), 2)
+        self.assertEqual(second.parent_self, first)
+
+
+class TestInheritModel(TestCase):
+    def test_inheritence_model(self):
+        filler = AutoFixture(InheritModel)
+        filler.create(10)
+        self.assertEqual(InheritModel.objects.count(), 10)
+
+    def test_inheritence_unique_together_model(self):
+        filler = AutoFixture(InheritUniqueTogetherModel)
+        filler.create(10)
+        self.assertEqual(InheritUniqueTogetherModel.objects.count(), 10)
+
 
 class TestUniqueConstraints(TestCase):
     def test_unique_field(self):
@@ -277,19 +325,19 @@ class TestGenerators(TestCase):
         result = generators.InstanceSelector(SimpleModel).generate()
         self.assertEqual(result.__class__, SimpleModel)
 
-        for i in xrange(10):
+        for i in range(10):
             result = generators.InstanceSelector(
                 SimpleModel, max_count=10).generate()
             self.assertTrue(0 <= len(result) <= 10)
             for obj in result:
                 self.assertEqual(obj.__class__, SimpleModel)
-        for i in xrange(10):
+        for i in range(10):
             result = generators.InstanceSelector(
                 SimpleModel, min_count=5, max_count=10).generate()
             self.assertTrue(5 <= len(result) <= 10)
             for obj in result:
                 self.assertEqual(obj.__class__, SimpleModel)
-        for i in xrange(10):
+        for i in range(10):
             result = generators.InstanceSelector(
                 SimpleModel, min_count=20, max_count=100).generate()
             # cannot return more instances than available
@@ -376,6 +424,14 @@ class TestRegistry(TestCase):
         self.assertTrue(SimpleModel in autofixture.REGISTRY)
         self.assertEqual(autofixture.REGISTRY[SimpleModel], SimpleAutoFixture)
 
+    def test_unregister(self):
+        autofixture.register(SimpleModel, SimpleAutoFixture)
+        self.assertTrue(SimpleModel in autofixture.REGISTRY)
+        self.assertEqual(autofixture.REGISTRY[SimpleModel], SimpleAutoFixture)
+
+        autofixture.unregister(SimpleModel)
+        self.assertFalse(SimpleModel in autofixture.REGISTRY)
+
     def test_create(self):
         autofixture.register(SimpleModel, SimpleAutoFixture)
         for obj in autofixture.create(SimpleModel, 10):
@@ -391,6 +447,32 @@ class TestRegistry(TestCase):
         obj = autofixture.create_one(
             SimpleModel, field_values={'name': 'bar'})
         self.assertEqual(obj.name, 'bar')
+
+    def test_registered_fixture_is_used_for_fk(self):
+        class BasicModelFixture(AutoFixture):
+            field_values={'chars': 'Hello World!'}
+
+        autofixture.register(BasicModel, BasicModelFixture)
+
+        fixture = AutoFixture(RelatedModel, generate_fk=['related'])
+        obj = fixture.create_one()
+        self.assertTrue(obj)
+        self.assertEqual(obj.related.chars, 'Hello World!')
+
+    def test_registered_fixture_is_used_for_m2m(self):
+        class SimpleModelFixture(AutoFixture):
+            field_values={'name': 'Jon Doe'}
+
+        autofixture.register(SimpleModel, SimpleModelFixture)
+
+        fixture = AutoFixture(M2MModel, generate_m2m={'m2m': (5,5)})
+        obj = fixture.create_one()
+        self.assertTrue(obj)
+
+        self.assertEqual(obj.m2m.count(), 5)
+        self.assertEqual(
+            list(obj.m2m.values_list('name', flat=True)),
+            ['Jon Doe'] * 5)
 
 
 class TestAutofixtureAPI(TestCase):
@@ -437,16 +519,16 @@ class TestManagementCommand(TestCase):
         self.command.handle(*models, **self.options)
         self.assertEqual(SimpleModel.objects.count(), 0)
 
-        models = ('autofixture_test.SimpleModel:1',)
+        models = ('autofixture_tests.SimpleModel:1',)
         self.command.handle(*models, **self.options)
         self.assertEqual(SimpleModel.objects.count(), 1)
 
-        models = ('autofixture_test.SimpleModel:5',)
+        models = ('autofixture_tests.SimpleModel:5',)
         self.command.handle(*models, **self.options)
         self.assertEqual(SimpleModel.objects.count(), 6)
 
     def test_generate_fk(self):
-        models = ('autofixture_test.DeepLinkModel2:1',)
+        models = ('autofixture_tests.DeepLinkModel2:1',)
         self.options['generate_fk'] = 'related,related__related'
         self.command.handle(*models, **self.options)
         obj = DeepLinkModel2.objects.get()
@@ -455,7 +537,7 @@ class TestManagementCommand(TestCase):
         self.assertEqual(obj.related.related2, obj.related.related)
 
     def test_generate_fk_with_no_follow(self):
-        models = ('autofixture_test.DeepLinkModel2:1',)
+        models = ('autofixture_tests.DeepLinkModel2:1',)
         self.options['generate_fk'] = 'related,related__related'
         self.options['no_follow_fk'] = True
         self.command.handle(*models, **self.options)
@@ -465,7 +547,7 @@ class TestManagementCommand(TestCase):
         self.assertEqual(obj.related.related2, None)
 
     def test_generate_fk_with_ALL(self):
-        models = ('autofixture_test.DeepLinkModel2:1',)
+        models = ('autofixture_tests.DeepLinkModel2:1',)
         self.options['generate_fk'] = 'ALL'
         self.command.handle(*models, **self.options)
         obj = DeepLinkModel2.objects.get()
@@ -477,7 +559,7 @@ class TestManagementCommand(TestCase):
     def test_no_follow_m2m(self):
         AutoFixture(SimpleModel).create(1)
 
-        models = ('autofixture_test.NullableFKModel:1',)
+        models = ('autofixture_tests.NullableFKModel:1',)
         self.options['no_follow_m2m'] = True
         self.command.handle(*models, **self.options)
         obj = NullableFKModel.objects.get()
@@ -487,7 +569,7 @@ class TestManagementCommand(TestCase):
         AutoFixture(SimpleModel).create(10)
         AutoFixture(OtherSimpleModel).create(10)
 
-        models = ('autofixture_test.M2MModel:25',)
+        models = ('autofixture_tests.M2MModel:25',)
         self.options['follow_m2m'] = 'm2m:3:3,secondm2m:0:10'
         self.command.handle(*models, **self.options)
 
@@ -496,7 +578,7 @@ class TestManagementCommand(TestCase):
             self.assertTrue(0 <= obj.secondm2m.count() <= 10)
 
     def test_generate_m2m(self):
-        models = ('autofixture_test.M2MModel:10',)
+        models = ('autofixture_tests.M2MModel:10',)
         self.options['generate_m2m'] = 'm2m:1:1,secondm2m:2:5'
         self.command.handle(*models, **self.options)
 
@@ -513,14 +595,47 @@ class TestManagementCommand(TestCase):
 
     def test_using_registry(self):
         autofixture.register(SimpleModel, SimpleAutoFixture)
-        models = ('autofixture_test.SimpleModel:10',)
+        models = ('autofixture_tests.SimpleModel:10',)
         self.command.handle(*models, **self.options)
         for obj in SimpleModel.objects.all():
             self.assertEqual(obj.name, 'foo')
 
     def test_use_option(self):
-        self.options['use'] = 'autofixture_tests.autofixture_test.tests.SimpleAutoFixture'
-        models = ('autofixture_test.SimpleModel:10',)
+        self.options['use'] = 'autofixture_tests.tests.SimpleAutoFixture'
+        models = ('autofixture_tests.SimpleModel:10',)
         self.command.handle(*models, **self.options)
         for obj in SimpleModel.objects.all():
             self.assertEqual(obj.name, 'foo')
+
+
+class TestGenericRelations(TestCase):
+    def assertNotRaises(self, exc_type, func, msg=None,
+            args=None, kwargs=None):
+        args = args or []
+        kwargs = kwargs or {}
+        try:
+            func(*args, **kwargs)
+        except exc_type as exc:
+            if msg is not None and exc.message != msg:
+                return
+            self.fail('{} failed with {}'.format(func, exc))
+
+    def test_process_gr(self):
+        """Tests the bug when GenericRelation field being processed
+        by autofixture.base.AutoFixtureBase#process_m2m
+        and through table appears as None.
+        """
+        count = 10
+        fixture = AutoFixture(GRModel)
+        self.assertNotRaises(AttributeError, fixture.create,
+            msg="'NoneType' object has no attribute '_meta'", args=[count])
+        self.assertEqual(GRModel.objects.count(), count)
+
+
+class TestShortcuts(TestCase):
+    def test_commit_kwarg(self):
+        instances = autofixture.create(BasicModel, 3, commit=False)
+        self.assertEqual([i.pk for i in instances], [None] * 3)
+
+        instance = autofixture.create_one(BasicModel, commit=False)
+        self.assertEqual(instance.pk, None)